테스트케이스 입력값 에러메세지 체크 함수 위치 수정
